requirements.txt
boto3
discord
marshmallow_dataclass